@@ -12,13 +12,13 @@
 	"devDependencies": {
 		"@microflash/remark-figure-caption": "^2.0.2",
 		"@sveltejs/adapter-static": "^3.0.0",
-		"@sveltejs/enhanced-img": "^0.3.4",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/enhanced-img": "^0.3.9",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
 		"@tailwindcss/forms": "^0.5.3",
 		"@tailwindcss/typography": "^0.5.9",
-		"@threlte/core": "^7.2.1",
-		"@threlte/extras": "^8.10.3",
+		"@threlte/core": "^8.0.1",
+		"@threlte/extras": "^9.0.1",
 		"@types/d3": "^7.4.3",
 		"@types/eslint": "^9.6.0",
 		"@types/three": "^0.168.0",
@@ -27,14 +27,14 @@
 		"camera-controls": "^2.8.3",
 		"d3": "^7.9.0",
 		"eslint": "^9.0.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
-		"heroicons-svelte": "^2.0.0",
+		"eslint-config-prettier": "^10.0.1",
+		"eslint-plugin-svelte": "^2.45.1",
+		"heroicons-svelte": "^2.0.2",
 		"mdsvex": "^0.12.3",
 		"mdsvex-relative-images": "^1.0.3",
 		"postcss": "^8.4.23",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"rehype-accessible-emojis": "^0.3.2",
 		"rehype-autolink-headings": "^7.1.0",
 		"rehype-slug": "^6.0.0",
@@ -42,21 +42,21 @@
 		"remark-gfm": "^4.0.0",
 		"remark-toc": "^9.0.0",
 		"remark-unwrap-images": "^4.0.0",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0",
 		"svelte-check": "^4.0.0",
 		"svelte-preprocess": "^6.0.2",
-		"svelte-youtube-embed": "^0.3.0",
+		"svelte-youtube-embed": "^0.4.4",
 		"sveltekit-embed": "^0.0.12",
 		"tailwindcss": "^3.3.1",
-		"three": "^0.168.0",
+		"three": "^0.172.0",
 		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
+		"typescript": "^5.5.0",
 		"typescript-eslint": "^8.0.0",
-		"vite": "^5.0.0",
+		"vite": "^6.0.11",
 		"vite-plugin-glsl": "^1.1.2"
 	},
 	"type": "module",
 	"dependencies": {
-		"eslint-plugin-svelte": "^2.32.4"
+		"eslint-plugin-svelte": "^2.45.1"
 	}
 }
