@@ -1,5 +1,9 @@
 <script lang="ts">
-	export let description: string;
+	interface Props {
+		description: string;
+	}
+
+	let { description }: Props = $props();
 </script>
 
 {#if description}
