@@ -6,10 +6,10 @@ declare global {
 		// interface Locals {}
 		// interface PageData {}
 		// interface Platform {}
-		interface MdsvexFile {
-			default: import('svelte/internal').SvelteComponent;
-			metadata: Record<string, string>;
-		}
+		// interface MdsvexFile {
+		// 	default: import('svelte/internal').SvelteComponent;
+		// 	metadata: Record<string, string>;
+		// }
 
 		type MdsvexResolver = () => Promise<MdsvexFile>;
 
