@@ -1,9 +1,13 @@
 <script lang="ts">
 	import WebGlCanvas from '$lib/components/WebGLCanvas.svelte';
-	export let fragShader: string;
-	export let webglVersion = 1;
+	interface Props {
+		fragShader: string;
+		webglVersion?: number;
+	}
 
-	let shaderCode = fragShader;
+	let { fragShader, webglVersion = 1 }: Props = $props();
+
+	let shaderCode = $state(fragShader);
 </script>
 
 {#key shaderCode}
@@ -12,5 +16,5 @@
 
 <details>
 	<summary>GLSL code (editable)</summary>
-	<textarea bind:value={shaderCode} class="w-full h-96 p-4 font-mono text-xs" />
+	<textarea bind:value={shaderCode} class="w-full h-96 p-4 font-mono text-xs"></textarea>
 </details>
