@@ -1,7 +1,8 @@
 import glsl from 'vite-plugin-glsl';
 import { sveltekit } from '@sveltejs/kit/vite';
+import { enhancedImages } from '@sveltejs/enhanced-img';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [glsl(), sveltekit()]
+	plugins: [enhancedImages(), glsl(), sveltekit()]
 });
